Add option to re-use HelmReleaseProxies
In order to update Clusters one-by-one, it is a normal strategy to
update the labels of a cluster. Even if there is another HelmChartProxy
that has a label selector that fits the new label of the Cluster, the
current process is to remove the HelmReleaseProxy and re-create it for
the other HelmChartProxy, which also means that the Helmchart is
uninstalled and installed again. This is, of course, not desirable, as
important applications might be affected.

Instead, HelmReleaseProxies can now be re-used if there is another
HelmChartProxy which references the same chart and which fits the new
labels of a Cluster.